Fixes #6 issue with source sync directory when :hostpath contains tilde expansion
When a legal source path in the Vagrantfile for synced directories contains tilde expansion (~/) shorthand for the current user's directory, notify-forwarder on the host fails with the following error:

libc++abi.dylib: terminating with uncaught exception of type std::runtime_error: Could not resolve path

This is because File.absolute_path() treats the tilde as a relative path to the current working directory, not the user's directory.  The current version of notify-forwarder seems to be able to deal with tilde expansion natively, so we can pass the :host path values directly to notify-forwarder.